AvatarMapper command dispatcher logic

+++ GoodVibes.Client.PiShock/Events/PiShockIntensityChangedEvent.cs
@@ -5,5 +5,5 @@ namespace GoodVibes.Client.PiShock.Events;
 public class PiShockIntensityChangedEvent : IEvent
 {
     public string? ToyId { get; set; }
-    public int Intensity { get; set; }
+    public float Intensity { get; set; }
 }
